[fix/#18] 홈 - 투표 플로팅 버튼 등등
