+++ app/src/main/java/com/sopt/sopkathon37android2/presentation/home/HomeViewModel.kt
@@ -21,33 +21,6 @@ class HomeViewModel @Inject constructor(
     private val _uiState = MutableStateFlow(HomeState())
     val uiState: StateFlow<HomeState> = _uiState.asStateFlow()
 
-    init {
-        val dummyVotes = listOf(
-            VoteUiModel(
-                id = 1L,
-                detailGroupTag = "총학생회",
-                tagType = TagType.MY,
-                studentCouncil = true,
-                voteTotalNumber = 123,
-                detailTitle = "중앙도서관 24시간 개방",
-                detailGroup = "시험기간 야간 학습 공간 확대"
-            ),
-            VoteUiModel(
-                id = 2L,
-                detailGroupTag = "단과대",
-                tagType = TagType.ALL,
-                studentCouncil = false,
-                voteTotalNumber = 87,
-                detailTitle = "휴게실 전자레인지 추가 설치",
-                detailGroup = "공학관 휴게실 환경 개선"
-            )
-        )
-
-        _uiState.value = _uiState.value.copy(
-            voteList = dummyVotes
-        )
-    }
-
     fun getDummy() {
         viewModelScope.launch {
             runCatching {
@@ -78,14 +51,6 @@ class HomeViewModel @Inject constructor(
                 selectedTab = HomeTab.VOTE,
                 isVoteSelected = true,
                 isIssueSelected = false
-            )
-        }
-    }
-
-    fun onToggleClicked() {
-        _uiState.update {
-            it.copy(
-                isActivated = !it.isActivated
             )
         }
     }
