+++ app/src/main/java/com/sopt/sopkathon37android2/presentation/home/VoteState.kt
@@ -1,0 +1,7 @@
+package com.sopt.sopkathon37android2.presentation.home
+
+import com.sopt.sopkathon37android2.presentation.home.model.VoteUiModel
+
+data class VoteState(
+    val voteList : List<VoteUiModel> = emptyList(),
+)
