+++ app/src/main/java/com/sopt/sopkathon37android2/presentation/home/navigation/HomeNavigation.kt
@@ -1,12 +1,14 @@
 package com.sopt.sopkathon37android2.presentation.home.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.sopt.sopkathon37android2.core.navigation.Route
 import com.sopt.sopkathon37android2.presentation.home.HomeRoute
+import com.sopt.sopkathon37android2.presentation.home.HomeViewModel
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateToHome(navOptions: NavOptions? = null) {
@@ -15,6 +17,7 @@ fun NavController.navigateToHome(navOptions: NavOptions? = null) {
 
 fun NavGraphBuilder.homeGraph(
     paddingValues: PaddingValues,
+    homeViewModel: HomeViewModel,
     onNavigateToRegister: () -> Unit,
     onNavigateToVote: () -> Unit,
 ) {
@@ -22,7 +25,8 @@ fun NavGraphBuilder.homeGraph(
         HomeRoute(
             paddingValues = paddingValues,
             onNavigateToRegister = onNavigateToRegister,
-            onNavigateToVote = onNavigateToVote
+            onNavigateToVote = onNavigateToVote,
+            viewModel = homeViewModel
         )
     }
 }
